python-nlp/requirements.txt
flask
flask-cors
pymongo
nltk
numpy
scikit-learn
spacy
pdfminer.six
python-dotenv
